Setup script to process list from schema directory
This change updates the setup script to read the SQL schema out of the
schema directory. It also changes the processing from chained psql calls
into a loop that processes a list of schema source files. This method is
more consistent and has been setup to correctly stop when an error is
encountered.